Create LICENSE
I chose GNU General Public License v3 because of what your project seems to be for, with a course/guide like yours, changes and improvements need to be made. This license allows you to do that while not letting anyone privately distribute it. Which is very good for projects in which many people are going to contribute.